webapp/pyproject.toml
[project]
name = "webapp"
version = "0.0.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ruff (>=0.11.0,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"

[tool.ruff]
ignore = ["F401"]
